V1.0/src/requirements.txt
glfw==2.1.0
h5py==3.6.0
numpy==1.21.4
Pillow==8.4.0
pygame==2.1.0
PyGLM==1.1.7
PyOpenGL==3.1.5
